+++ packages/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "@ai-job-interview/packages",
+  "version": "0.1.0",
+  "private": true,
+  "main": "./index.ts",
+  "types": "./index.ts",
+  "scripts": {
+    "build": "tsc",
+    "type-check": "tsc --noEmit"
+  },
+  "dependencies": {
+    "zod": "^3.25.23"
+  },
+  "devDependencies": {
+    "typescript": "^5.6.0"
+  }
+}
