+++ pnpm-workspace.yaml
@@ -1,3 +1,3 @@
 packages:
   - "apps/*"
-  - "packages/*"
+  - "packages"
